refactor: give request-body parsing one module
`parseJsonBody` moves from `routes/shared.ts` into `routes/body.ts`, and
`parseBody` is composed on it, so the raw reader and the schema-backed
form share one try/catch. `parseBody` awaits `safeParseAsync` so a schema
with an async refinement answers 400 instead of escaping as a 500.
