+++ packages/control-plane/src/routes/body.ts
@@ -10,6 +10,20 @@ export function bodyIssue(failure: z.ZodError): string {
 }
 
 /**
+ * Read the request body as JSON, or answer the route's 400.
+ *
+ * For routes that validate the raw value themselves; `parseBody` is the
+ * schema-backed form.
+ */
+export async function parseJsonBody<T>(request: Request): Promise<T | Response> {
+  try {
+    return (await request.json()) as T;
+  } catch {
+    return error("Invalid JSON body", 400);
+  }
+}
+
+/**
  * Parse a request's JSON body with `schema`, or answer the route's 400.
  *
  * A body that is not JSON answers `Invalid JSON body`. A schema failure
@@ -21,13 +35,9 @@ export async function parseBody<Schema extends z.ZodType>(
   schema: Schema,
   message?: string
 ): Promise<z.output<Schema> | Response> {
-  let raw: unknown;
-  try {
-    raw = await request.json();
-  } catch {
-    return error("Invalid JSON body", 400);
-  }
-  const result = schema.safeParse(raw);
+  const raw = await parseJsonBody<unknown>(request);
+  if (raw instanceof Response) return raw;
+  const result = await schema.safeParseAsync(raw);
   if (!result.success) return error(message ?? bodyIssue(result.error), 400);
   return result.data;
 }
