+++ packages/control-plane/src/routes/integration-settings.ts
@@ -33,9 +33,9 @@ import {
   GITHUB_USER_OR_SERVICE_ROUTE,
   json,
   error,
-  parseJsonBody,
   requirePermission,
 } from "./shared";
+import { parseJsonBody } from "./body";
 
 const logger = createLogger("router:integration-settings");
